chore(riceType): change price to integer and update tests; remove duplicate e2e spec; update seed for Secado

+++ libs/entities/rice-type.entity.ts
@@ -22,7 +22,7 @@ export class RiceType {
   @Column({ nullable: true })
   description: string;
 
-  @Column('decimal')
+  @Column('int')
   price: number;
 
   @Column()
